Generador de Reportes
la funcionalidad de "Generar Reporte" ya se encuentra disponible, se comentaron varias lineas de codigo y se solucionaron algunos bugs al querer ver el historial de una cuenta

+++ src/Controller/ControladorDepositos.java
@@ -40,7 +40,7 @@ public class ControladorDepositos implements ActionListener{
     
     //accion que realizara el boton Aceptar
     public void actionPerformed(ActionEvent e) {
-        float efectivo= Integer.parseInt(view.CajaMonto.getText());
+        float efectivo= Float.parseFloat(view.CajaMonto.getText());
         int contador = view.CajaCuenta.getSelectedIndex();
         
         if(efectivo>0 && cuentas.get(contador).getTransaccion().size()<=24){
